+++ ml_tools/ML_datasetmaster/_base_datasetmaster.py
@@ -684,7 +684,7 @@ class _BaseDatasetMaker(ABC):
         
         if verbose:
             _LOGGER.info(
-                f"Dataset loaded from '{target_filepath.name}' with ID '{instance.id}'.\n"
+                f"Dataset loaded from '{target_filepath.name}'.\n"
                 f"{repr(instance)}"
             )