+++ ml_tools/ML_datasetmaster/_base_sequence_dataset.py
@@ -527,7 +527,7 @@ class _BaseSequenceDataset(_BaseDatasetMaker, ABC):
         
         if verbose:
             _LOGGER.info(
-                f"Dataset loaded from '{target_filepath.name}' with ID '{instance.id}'.\n"
+                f"Dataset loaded from '{target_filepath.name}'.\n"
                 f"{repr(instance)}"
             )